requirements.txt
fastapi
uvicorn
pyyaml
requests
python-multipart
